+++ src/api/readDevice/tests/test_index.py
@@ -3,11 +3,13 @@ import os
 from unittest import mock
 
 import pytest
+from domain.core.device_key import DeviceKeyType
 from domain.core.root import Root
 from domain.repository.device_repository import DeviceRepository
 from nhs_context_logging import app_logger
 
 from test_helpers.dynamodb import mock_table
+from test_helpers.response_assertions import _response_assertions
 from test_helpers.uuid import consistent_uuid
 
 TABLE_NAME = "hiya"
@@ -20,15 +22,13 @@ TABLE_NAME = "hiya"
     ],
 )
 def test_index(version):
-    device_id = "XXX-YYY"
+    device_key = "P.XXX-YYY"
     org = Root.create_ods_organisation(ods_code="ABC")
     product_team = org.create_product_team(
         id=consistent_uuid(1), name="product-team-name"
     )
-    device = product_team.create_device(
-        id=device_id, name="device-name", type="product"
-    )
-    device.add_key(type="product_id", key=device_id)
+    device = product_team.create_device(name="device-name", type="product")
+    device.add_key(DeviceKeyType.PRODUCT_ID, device_key)
 
     with mock_table(TABLE_NAME) as client, mock.patch.dict(
         os.environ,
@@ -46,7 +46,7 @@ def test_index(version):
         result = handler(
             event={
                 "headers": {"version": version},
-                "pathParameters": {"id": device_id},
+                "pathParameters": {"id": str(device.id)},
             }
         )
 
@@ -61,7 +61,7 @@ def test_index(version):
                 }
             },
             "identifier": [
-                {"system": "connecting-party-manager/product_id", "value": "XXX-YYY"}
+                {"system": "connecting-party-manager/product_id", "value": "P.XXX-YYY"}
             ],
             "owner": {
                 "identifier": {
@@ -72,15 +72,19 @@ def test_index(version):
         }
     )
 
-    assert result == {
+    expected = {
         "statusCode": 200,
         "body": expected_result,
         "headers": {
             "Content-Length": str(len(expected_result)),
             "Content-Type": "application/json",
             "Version": version,
+            "Location": None,
         },
     }
+    _response_assertions(
+        result=result, expected=expected, check_body=True, check_content_length=True
+    )
 
 
 @pytest.mark.parametrize(
@@ -135,12 +139,16 @@ def test_index_no_such_device(version):
         }
     )
 
-    assert result == {
+    expected = {
         "statusCode": 404,
         "body": expected_result,
         "headers": {
             "Content-Length": str(len(expected_result)),
             "Content-Type": "application/json",
             "Version": version,
+            "Location": None,
         },
     }
+    _response_assertions(
+        result=result, expected=expected, check_body=True, check_content_length=True
+    )